modified ft_atoi.c

+++ ft_atoi.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   ft_atoi.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: koimai <koimai@student.42.fr>              +#+  +:+       +#+        */
+/*   By: koimai <koimai@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/21 12:09:45 by koimai            #+#    #+#             */
-/*   Updated: 2023/09/30 12:35:54 by koimai           ###   ########.fr       */
+/*   Updated: 2023/10/06 09:29:30 by koimai           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	check_long(int flag, long num, int digit)
+int	check_long(int flag, long long num, int digit)
 {
 	if (flag == 1)
 	{
@@ -31,7 +31,7 @@ int	check_long(int flag, long num, int digit)
 
 int	ft_atoi(const char *str)
 {
-	long	result;
+	long long	result;
 	int		flag;
 	int		i;
